Working Code for Base Station
Many of the functions that were not needed were deleted. This code that
is uploaded in the commit works on the base station. It is able to
receive data from the remote station and display the data on the screen.
There is also a humidity and temperature sensor on this base station
that is also pulling data.